updated userHomePage component to match Home component. Created commit to buy modal displaying on both userHomePage and Home components
