+++ src/components/UserHomePage.jsx
@@ -10,9 +10,9 @@ import Header from './Header.jsx';
 import ImageHeader from './ImageHeader.jsx';
 import Results from './Results.jsx';
 import Search from './Search.jsx';
-// import SaveEvent from './SaveEvent.jsx';
 import { connect } from 'react-redux';
 import helpers from '../actions/helpers.js';
+import Commit from './Commit.jsx';
 import Footer from './Footer.jsx';
 
 class UserHomePage extends React.Component {
@@ -35,24 +35,6 @@ class UserHomePage extends React.Component {
     return this.setState({ searchResults: results});
   }
 
-  // T&C to commit to purchase
-  displayModal() {
-    let modal = document.getElementById('commitModal');
-    let btn = document.querySelector("buy");
-    modal.style.display = "block";
-    window.onclick = (event) => {
-      if (event.target == modal) {
-        modal.style.display = "none";
-      }
-    }
-  }
-
-  closeModal() {
-    let modal = document.getElementById('commitModal');
-    let span = document.querySelector("close");
-    modal.style.display = "none";
-  }
-
   render() {
     const location = {
         lat: 34.0523003,
@@ -66,48 +48,29 @@ class UserHomePage extends React.Component {
         <Header />
         <About />
         <ImageHeader />
-        <SaveEvent />
         <Search setSearchResults={this.setSearchResults} />    
-
-        {/*section for display search results*/}
-        <div className="home-nav row">
-          Search results
-        </div>
-        <br/>
-        <br/>
-        <br/>
-        <div className="row">
-          <div className="col-md-1"></div>
-          <div className="col-md-6">
+  
             <div className="event-results">
               <Results searchResults={this.state.searchResults}/>
             </div>
-          </div>
+     
 
-          {/*place holder for displaying map*/}
-          <div className="col-md-3">
+
             <div className = "mapAPI">
-                <div style={{width:600, height:400}}>
+                <div style={{width: '100%', height:700}}>
                   <Map center={location} events={this.state.searchResults} />
                 </div>
-            </div>
-          </div>
-        </div>
+                </div>
+       
         <Footer />
 
-
-          {/*<div className="buy" onClick={this.displayModal}>Commit to buy</div>*/}
-          {/*place holder for displaying map*/}
-          {/*<div className = "mapAPI">
-            Map goes here
-          </div>
-
+          {/*modal to commit to buy*/}
           <div id="commitModal" className="modal">
             <div className="modal-content">
               <span className="close" onClick={this.closeModal}>&times;</span>
               <Commit />
             </div>
-          </div>*/}
+          </div>
 
       </div>
     );
